start of a typescript conversion
