+++ .gitattributes
@@ -1,1 +1,1 @@
-*.min.js -diff
+dist/* -diff
