+++ index.ts
@@ -1,0 +1,5 @@
+/// <reference path="./typings/node/node.d.ts"/>
+/// <reference path="./lib/index"/>
+
+"use strict";
+module.exports = Welsh;
