+++ test/helpers.js
@@ -3,7 +3,8 @@
 "use strict";
 
 var expect = require('chai').expect;
-var helpers = require('../lib/helpers');
+var welsh = require('../dist/welsh-node');
+var helpers = welsh.helpers;
 
 describe("Welsh Helper tryCatch", function () {
   it("should allow Try block", function (done) {
